refactor: consolidate analyze/cli.py into cli/analyze.py
Move all orchestration logic (run_typed_analysis, save_results,
print_summary, data loading, analyzer creation) from the separate
analyze/cli.py into the existing cli/analyze.py entry point.

- Delete src/oumi/analyze/cli.py and its duplicate ANALYZER_REGISTRY
- Use core registry (REGISTRY.get_sample_analyzer) instead
- Update analyze/__init__.py to re-export from oumi.cli.analyze
- Update tests to use core registry for registration checks
